feat(twitch): improvement deleting, rendering channels without template

+++ twitch/views.py
@@ -1,14 +1,20 @@
+import json
+from django.http import JsonResponse
 from django.shortcuts import render, redirect
-
 from .models import Twitch_model
 
 
-def get_streamer(request):
-    streamers = list(Twitch_model.objects.all().values("streamer", "id"))
-    return render(request, "twitch/twitch.html", {"streamers": streamers})
+def streamers_view(request):
+    return render(request, "twitch/twitch.html")
 
+def streamers_api(request):
+    if request.method == 'GET':
+        streamers = list(Twitch_model.objects.all().values("streamer", "id"))
+        return JsonResponse({"ok": True, "streamers": streamers})
 
-def delete_streamer(request, pk):
-    s = Twitch_model.objects.get(pk=pk)
-    s.delete()
-    return redirect("twitch:twitch")
+    if request.method == 'DELETE':
+        body = json.loads(request.body)
+        channel_id = body["id"]
+        s = Twitch_model.objects.get(pk=channel_id)
+        s.delete()
+        return JsonResponse({"ok": True})
